--- lib/commons/view/splash_screen.dart
@@ -1,64 +1,0 @@
-import 'package:bookbox/commons/const/size.dart';
-import 'package:bookbox/commons/view/root_tab.dart';
-import 'package:bookbox/home/view/home_screen.dart';
-import 'package:bookbox/library/view/library_screen.dart';
-import 'package:flutter/material.dart';
-
-class SplashPage extends StatefulWidget {
-  @override
-  _SplashPageState createState() => _SplashPageState();
-}
-
-class _SplashPageState extends State<SplashPage>
-    with SingleTickerProviderStateMixin {
-  double _opacity = 0.0;
-
-  @override
-  void initState() {
-    super.initState();
-    // 일정 시간 후에 애니메이션 시작
-    Future.delayed(Duration(milliseconds: 500), () {
-      setState(() {
-        _opacity = 1.0;
-      });
-    });
-
-    // 3초 후에 메인 페이지로 이동
-    Future.delayed(Duration(seconds: 3), () {
-      Navigator.pushReplacement(
-        context,
-        MaterialPageRoute(builder: (context) => RootTab()),
-      );
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Center(
-        child: AnimatedOpacity(
-          opacity: _opacity,
-          duration: Duration(seconds: 2),
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              Image.asset(
-                "assets/logo.png",
-                width: 200,
-                height: 200,
-              ),
-              SizedBox(height: gap_s),
-              Text(
-                "BookBox",
-                style: TextStyle(
-                    fontSize: 40.0,
-                    color: Colors.black,
-                    fontWeight: FontWeight.bold),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
