+++ lib/ui/main/home/cate_tab/cate_tab.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+class CateTab extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      color: Colors.blue[100],
+      child: Center(
+        child: Text(
+          '주제별',
+          style: TextStyle(fontSize: 20, color: Colors.black),
+        ),
+      ),
+    );
+  }
+}
