+++ lib/ui/main/home/hit_tab/hit_tab.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+class HitTab extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      color: Colors.blue[100],
+      child: Center(
+        child: Text(
+          '내서재',
+          style: TextStyle(fontSize: 20, color: Colors.black),
+        ),
+      ),
+    );
+  }
+}
