+++ lib/ui/main/home/recommend_tab/recommend_tab.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+class RecommendTab extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      color: Colors.orange[100],
+      child: Center(
+        child: Text(
+          '내서재',
+          style: TextStyle(fontSize: 20, color: Colors.black),
+        ),
+      ),
+    );
+  }
+}
