+++ lib/ui/main/home/uptodate_tab/uptodate_tab.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+class UptodateTab extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      color: Colors.blue[100],
+      child: Center(
+        child: Text(
+          '내서재',
+          style: TextStyle(fontSize: 20, color: Colors.black),
+        ),
+      ),
+    );
+  }
+}
