+++ lib/ui/main/library/history_tab/history_tab.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+class HistoryTab extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      color: Colors.green[100],
+      child: Center(
+        child: Text(
+          '서재탭4',
+          style: TextStyle(fontSize: 20, color: Colors.black),
+        ),
+      ),
+    );
+  }
+}
