+++ lib/ui/main/library/lend_tab/lend_tab.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+class LendTab extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      color: Colors.orange[100],
+      child: Center(
+        child: Text(
+          '서재탭1',
+          style: TextStyle(fontSize: 20, color: Colors.black),
+        ),
+      ),
+    );
+  }
+}
