+++ lib/ui/main/library/reserve_tab/reserve_tab.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+class ReserveTab extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      color: Colors.blue[100],
+      child: Center(
+        child: Text(
+          '서재탭 2',
+          style: TextStyle(fontSize: 20, color: Colors.black),
+        ),
+      ),
+    );
+  }
+}
