+++ lib/ui/main/library/review_tab/review_tab.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+class ReviewTab extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      color: Colors.yellowAccent[100],
+      child: Center(
+        child: Text(
+          '서재탭3',
+          style: TextStyle(fontSize: 20, color: Colors.black),
+        ),
+      ),
+    );
+  }
+}
