+++ pubspec.yaml
@@ -36,6 +36,8 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.8
   google_fonts: ^6.2.1
+  dio: ^5.5.0+1
+  flutter_riverpod: ^2.5.1
 
 dev_dependencies:
   flutter_test:
